Moving window initial implement
When MOVING_WINDOW=1 is specified, a window moves with the
visibilities. Each thread computes the contribution from
one point in the GCF, as in the first gather implementation.
The difference here is that threads will only ever use image
points on a square grid with pitch GCF_DIM and offset by
threadIdx.x, threadIdx.y.